Fix sticky nav
There's a typo in foundation.min.js that prevents the sticky nav from working. 
See here: https://github.com/zurb/foundation/issues/3608#issuecomment-29071058
This is a known issue and may have been fixed in foundation by now but for the time being this'll get it working in Reverie.
I've tested and can confirm that this fixes the problem.